refactor: extract TabBadge into its own file
